Add files via upload
This NFC mobile tag is constructed from durable hard PET and integrates the NXP® NTAG®213 chip. With a diameter of 30mm, it is optimized for on-metal performance. It supports ISO/IEC 14443-A standards and is ideal for mobile phone attachment, asset tracking, and business card sharing.
Product link: https://rfidtag.com/product/on-metal-nfc-hard-pet-mobile-tag-nxp-ntag-213/